[625] Amended application controller to be skinny
