+++ app/controllers/api/public/v1/courses_controller.rb
@@ -2,6 +2,7 @@ module API
   module Public
     module V1
       class CoursesController < API::Public::V1::ApplicationController
+        include PagyPagination
         def index
           render jsonapi: paginated_records,
                  include: include_param,
